发布资源清单下载、高级搜索、合规趋势统计等API.

+++ aliyun-java-sdk-config/src/main/java/com/aliyuncs/config/model/v20200907/ListAggregateResourceRelationsResponse.java
@@ -89,7 +89,7 @@ public class ListAggregateResourceRelationsResponse extends AcsResponse {
 
 		public static class ResourceRelationListItem {
 
-			private Integer accountId;
+			private Long accountId;
 
 			private String sourceResourceType;
 
@@ -103,11 +103,11 @@ public class ListAggregateResourceRelationsResponse extends AcsResponse {
 
 			private String relationType;
 
-			public Integer getAccountId() {
+			public Long getAccountId() {
 				return this.accountId;
 			}
 
-			public void setAccountId(Integer accountId) {
+			public void setAccountId(Long accountId) {
 				this.accountId = accountId;
 			}
